[SYCL] Deprecate ext::oneapi::sub_group
sub_group extension is deprecated in favor of standard SYCL2020
sub_group.
