+++ sycl/include/sycl/detail/type_traits.hpp
@@ -28,6 +28,7 @@ inline constexpr bool is_fixed_size_group_v = is_fixed_size_group<T>::value;
 } // namespace detail
 
 template <int Dimensions> class group;
+struct sub_group;
 namespace ext::oneapi {
 struct sub_group;
 
@@ -50,6 +51,7 @@ struct is_fixed_topology_group<sycl::group<Dimensions>> : std::true_type {};
 template <>
 struct is_fixed_topology_group<sycl::ext::oneapi::sub_group> : std::true_type {
 };
+template <> struct is_fixed_topology_group<sycl::sub_group> : std::true_type {};
 
 template <class T> struct is_user_constructed_group : std::false_type {};
 
@@ -77,6 +79,7 @@ struct is_group<group<Dimensions>> : std::true_type {};
 template <typename T> struct is_sub_group : std::false_type {};
 
 template <> struct is_sub_group<ext::oneapi::sub_group> : std::true_type {};
+template <> struct is_sub_group<sycl::sub_group> : std::true_type {};
 
 template <typename T>
 struct is_generic_group
