+++ sycl/include/sycl/ext/oneapi/experimental/opportunistic_group.hpp
@@ -8,6 +8,7 @@
 
 #pragma once
 #include <sycl/ext/oneapi/experimental/non_uniform_groups.hpp>
+#include <sycl/ext/oneapi/sub_group.hpp>
 #include <sycl/ext/oneapi/sub_group_mask.hpp>
 
 namespace sycl {
@@ -128,7 +129,7 @@ inline opportunistic_group get_opportunistic_group() {
 #ifdef __SYCL_DEVICE_ONLY__
 #if defined(__SPIR__)
   // TODO: It may be wiser to call the intrinsic than rely on this_group()
-  sycl::sub_group sg = sycl::ext::oneapi::this_sub_group();
+  sycl::sub_group sg = sycl::ext::oneapi::experimental::this_sub_group();
   sub_group_mask mask = sycl::ext::oneapi::group_ballot(sg, true);
   return opportunistic_group(mask);
 #elif defined(__NVPTX__)
