+++ sycl/include/sycl/ext/oneapi/experimental/uniform.hpp
@@ -22,19 +22,15 @@
 // Forward declarations of types not allowed to be wrapped in uniform:
 namespace sycl {
 __SYCL_INLINE_VER_NAMESPACE(_V1) {
-namespace ext::oneapi {
 
 struct sub_group;
-
-} // namespace ext::oneapi
-
 template <int, bool> class item;
 template <int> class id;
 template <int> class nd_item;
 template <int> class h_item;
 template <int> class group;
 template <int> class nd_range;
-using ext::oneapi::sub_group;
+using sycl::sub_group;
 
 } // __SYCL_INLINE_VER_NAMESPACE(_V1)
 } // namespace sycl
