+++ sycl/include/sycl/ext/oneapi/sub_group_mask.hpp
@@ -40,7 +40,8 @@ namespace ext::oneapi {
 // need to forward declare sub_group_mask first
 struct sub_group_mask;
 template <typename Group>
-std::enable_if_t<std::is_same_v<std::decay_t<Group>, sub_group>, sub_group_mask>
+std::enable_if_t<std::is_same_v<std::decay_t<Group>, sycl::sub_group>,
+                 sub_group_mask>
 group_ballot(Group g, bool predicate = true);
 
 struct sub_group_mask {
@@ -293,7 +294,8 @@ private:
 };
 
 template <typename Group>
-std::enable_if_t<std::is_same_v<std::decay_t<Group>, sub_group>, sub_group_mask>
+std::enable_if_t<std::is_same_v<std::decay_t<Group>, sycl::sub_group>,
+                 sub_group_mask>
 group_ballot(Group g, bool predicate) {
   (void)g;
 #ifdef __SYCL_DEVICE_ONLY__
