+++ sycl/test-e2e/GroupAlgorithm/SYCL2020/permute_select.hpp
@@ -29,7 +29,7 @@ void check(queue &Queue, size_t G = 256, size_t L = 64) {
       accessor acc_xor{buf_xor, cgh, sycl::read_write};
       accessor sgsizeacc{sgsizebuf, cgh, sycl::read_write};
       cgh.parallel_for<sycl_subgr<T, N>>(NdRange, [=](nd_item<1> NdItem) {
-        ext::oneapi::sub_group SG = NdItem.get_sub_group();
+        sycl::sub_group SG = NdItem.get_sub_group();
         uint32_t wggid = NdItem.get_global_id(0);
         uint32_t sgid = SG.get_group_id().get(0);
         vec<T, N> vwggid(wggid), vsgid(sgid);
@@ -90,7 +90,7 @@ template <typename T> void check(queue &Queue, size_t G = 256, size_t L = 64) {
       accessor acc_xor{buf_xor, cgh, sycl::read_write};
       accessor sgsizeacc{sgsizebuf, cgh, sycl::read_write};
       cgh.parallel_for<sycl_subgr<T, 0>>(NdRange, [=](nd_item<1> NdItem) {
-        ext::oneapi::sub_group SG = NdItem.get_sub_group();
+        sycl::sub_group SG = NdItem.get_sub_group();
         uint32_t wggid = NdItem.get_global_id(0);
         uint32_t sgid = SG.get_group_id().get(0);
         if (wggid == 0)
