+++ sycl/test-e2e/Matrix/get_coord_bf16_matA_impl.hpp
@@ -110,7 +110,7 @@ void matrix_sum_rows(queue q, big_matrix<T, M, K> &A, nd_range<2> &r) {
            const auto sg_startx = global_idx - spmd_item.get_local_id(0);
            const auto sg_starty = global_idy - spmd_item.get_local_id(1);
 
-           ext::oneapi::sub_group sg = spmd_item.get_sub_group();
+           sycl::sub_group sg = spmd_item.get_sub_group();
 
            // TM = 8, TK = 32
           joint_matrix<sub_group, int8_t, use::a, TM, TK, layout::row_major>
