+++ sycl/test-e2e/Matrix/get_coord_bf16_matB_impl.hpp
@@ -132,7 +132,7 @@ void matrix_sum_cols(queue q, big_matrix<T, M, N> &B, nd_range<2> &r) {
            const auto sg_startx = global_idx - spmd_item.get_local_id(0);
            const auto sg_starty = global_idy - spmd_item.get_local_id(1);
 
-           ext::oneapi::sub_group sg = spmd_item.get_sub_group();
+           sycl::sub_group sg = spmd_item.get_sub_group();
 
            // TK = 32, TN = 16
            joint_matrix<sub_group, int8_t, use::b, TK, TN,
