+++ sycl/test-e2e/SubGroup/generic-shuffle.cpp
@@ -37,7 +37,7 @@ void check_pointer(queue &Queue, size_t G = 256, size_t L = 64) {
 
       cgh.parallel_for<SpecializationKernelName>(
           NdRange, [=](nd_item<1> NdItem) {
-            ext::oneapi::sub_group SG = NdItem.get_sub_group();
+            sycl::sub_group SG = NdItem.get_sub_group();
             uint32_t wggid = NdItem.get_global_id(0);
             uint32_t sgid = SG.get_group_id().get(0);
             if (wggid == 0)
@@ -139,7 +139,7 @@ void check_struct(queue &Queue, Generator &Gen, size_t G = 256, size_t L = 64) {
 
       cgh.parallel_for<SpecializationKernelName>(
           NdRange, [=](nd_item<1> NdItem) {
-            ext::oneapi::sub_group SG = NdItem.get_sub_group();
+            sycl::sub_group SG = NdItem.get_sub_group();
             uint32_t wggid = NdItem.get_global_id(0);
             uint32_t sgid = SG.get_group_id().get(0);
             if (wggid == 0)
